requirements.txt
pillow>=9.0.0
tensorflow>=2.0.0
numpy>=1.19.0
